Last changes, including os2fs sources for Genode.

+++ OS2/Server/include/os3/ixfmgr.h
@@ -81,7 +81,7 @@ typedef
   {
     fnLoad * Load;            // Fill format structure
     fnFixup * Fixup;          // Fixup module address
-    unsigned long area;       // memory area identifier (for example, shared memory area)
+    unsigned long long area;  // memory area identifier (for example, shared memory area)
     void *addr;               // Load address
     unsigned long size;       // File size
     char *name;               // Fully-qualified filename
