+++ OS2/Server/include/os3/l4_alloc_mem.h
@@ -10,7 +10,7 @@
 
 void l4_test_mem_alloc(void);
 //void * l4_alloc_mem(unsigned long area, int base, int size, int flags, unsigned long PIC, l4dm_dataspace_t *ds);
-void * l4_alloc_mem(unsigned long area, int base, int size, int flags, unsigned long PIC, void *ds);
+void * l4_alloc_mem(unsigned long long area, int base, int size, int flags, unsigned long PIC, void *ds);
 int l4_translate_os2_flags(int flags);
 
 #ifdef __cplusplus
